New assets, component & script changes

+++ components/assets/shrinkImages.js
@@ -11,19 +11,21 @@ const fileList = fs.readdirSync(`../../public/images/`);
 
     for (let k = 0; k < folderImages.length; k++) {
       const currentImageName = folderImages[k];
-      const currentNum = currentImageName.split(".")[0];
+      if (currentImageName.includes(".png")) {
+        const currentNum = currentImageName.split(".")[0];
 
-      await sharp(`../../public/images/${file}/${currentImageName}`)
-            .toFile(`../../public/images/${file}/${currentNum}.webp`);
+        await sharp(`../../public/images/${file}/${currentImageName}`)
+              .toFile(`../../public/images/${file}/${currentNum}.webp`);
 
-      await sharp(`../../public/images/${file}/${currentImageName}`)
-            .resize(128, 128)
-            .toFile(`../../public/images/${file}/${currentNum}_thumb.webp`);
+        await sharp(`../../public/images/${file}/${currentImageName}`)
+              .resize(128, 128)
+              .toFile(`../../public/images/${file}/${currentNum}_thumb.webp`);
 
-      try {
-        fs.unlinkSync(`../../public/images/${file}/${currentImageName}`);
-      } catch (err) {
-        console.log(err);
+        try {
+          fs.unlinkSync(`../../public/images/${file}/${currentImageName}`);
+        } catch (err) {
+          console.log(err);
+        }
       }
     }
